reflect/requirements.txt
requests==2.20.0
google-cloud-firestore==1.6.0
